Code optimization 2
1. Created and added all scripts to Platform2D namespace
2. Replaced tags with custom Enum Tags.
3. Merged code from multiple scripts into one to reduce complexity.
4. Removed unused scripts.
