+++ Assets/Scripts/Collectables.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 900adb8bf5c8e0e468dae8a7b3e47191
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
