--- Assets/Scripts/Enums/Level.cs
@@ -1,10 +1,0 @@
-
-public enum Level
-{
-    LOBBY,
-    LEVEL1,
-    LEVEL2,
-    LEVEL3,
-    LEVEL4,
-    LEVEL5
-}
