--- Assets/Scripts/Enums/LevelStatus.cs
@@ -1,7 +1,0 @@
-
-public enum LevelStatus
-{
-    LOCKED,
-    UNLOCKED,
-    COMPLETED
-}
