+++ Assets/Scripts/Level/LevelLoader.cs
@@ -1,0 +1,30 @@
+using UnityEngine;
+using UnityEngine.UI;
+
+namespace Platatformer2D
+{
+    public class LevelLoader : MonoBehaviour
+    {
+        Button button;
+        [SerializeField] Level level;
+
+        void Awake()
+        {
+            button = GetComponent<Button>();
+            button.onClick.AddListener(LoadSelectedLevel);
+        }
+
+        void Start()
+        {
+            if (LevelManager.Instance.GetLevelStatus(level) == LevelStatus.LOCKED)
+                button.interactable = false;
+        }
+
+        void LoadSelectedLevel()
+        {
+            LevelStatus levelStatus = LevelManager.Instance.GetLevelStatus(level);
+            Utils.LoadLevel((int)level);
+        }
+    }
+}
+
