--- Assets/Scripts/Player/FallDetect.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class FallDetect : MonoBehaviour
-{
-    private void OnTriggerEnter2D(Collider2D collision)
-    {
-        collision.GetComponent<PlayerController>().KillPlayer();
-    }
-}
