--- Assets/Scripts/Player/LevelExit.cs
@@ -1,13 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class LevelExit : MonoBehaviour
-{
-    [SerializeField] LevelCompletedController levelCompletedController;
-    private void OnTriggerEnter2D(Collider2D collision)
-    {
-        LevelManager.Instance.LevelCompleted();
-        levelCompletedController.showLevelCompletedScreen();
-    }
-}
