+++ Assets/Scripts/UI/GameUIManager.cs
@@ -1,0 +1,23 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Platatformer2D
+{
+    public class GameUIManager : MonoBehaviour
+    {
+        GameObject gameOverPanel;
+        GameObject levelCompletedPanel;
+
+        void showGameOverPanel()
+        {
+            gameOverPanel.SetActive(true);
+        }
+
+        void showLevelCompletedPanel()
+        {
+            levelCompletedPanel.SetActive(true);
+        }
+    }
+}
+
