+++ Assets/Scripts/UI/GameUIManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b4784e93a2e73d64ea0490f0924aa88c
+guid: ed74562674423714da634f479abb759d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
