+++ Assets/Scripts/UI/UIButtonController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f8c1691d43c5fd843833fd928698f8c4
+guid: d8ea35974501c5f40b24fe8807f9dda5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
