+++ Assets/Scripts/Utility/MainEnums.cs
@@ -1,0 +1,38 @@
+﻿using System.Collections;
+using UnityEngine;
+
+namespace Platatformer2D
+{
+    public enum EnumTag
+    {
+        GROUND,
+        LEVEL_EXIT,
+        PIT
+    }
+
+    public enum Level
+    {
+        LOBBY,
+        LEVEL1,
+        LEVEL2,
+        LEVEL3,
+        LEVEL4,
+        LEVEL5
+    }
+
+    public enum LevelStatus
+    {
+        LOCKED,
+        UNLOCKED,
+        COMPLETED
+    }
+
+    public enum ButtonType
+    {
+        START,
+        EXIT,
+        RESTART,
+        QUIT,
+        NEXT_LEVEL
+    }
+}
