+++ UserSettings/EditorUserSettings.asset
@@ -15,22 +15,22 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c78111b122b29292623707c67083debf42d
       flags: 0
     RecentlyUsedScenePath-3:
-      value: 22424703114646680e0b0227036c73150012146a7d66333e243d04
+      value: 22424703114646680e0b0227036c73150012146a7966333e243d04
       flags: 0
     RecentlyUsedScenePath-4:
-      value: 22424703114646680e0b0227036c73150012146a7e66333e243d04
+      value: 22424703114646680e0b0227036c73150012146a7866333e243d04
       flags: 0
     RecentlyUsedScenePath-5:
       value: 22424703114646680e0b0227036c73150012146a7f66333e243d04
       flags: 0
     RecentlyUsedScenePath-6:
-      value: 22424703114646680e0b0227036c73150012146a7866333e243d04
+      value: 22424703114646680e0b0227036c73150012146a7e66333e243d04
       flags: 0
     RecentlyUsedScenePath-7:
-      value: 22424703114646680e0b0227036c73150012146a7966333e243d04
+      value: 22424703114646680e0b0227036c731f1415016439262f2434
       flags: 0
     RecentlyUsedScenePath-8:
-      value: 22424703114646680e0b0227036c731f1415016439262f2434
+      value: 22424703114646680e0b0227036c73150012146a7d66333e243d04
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
